Week09: Final Project Initial Changes
Created an initial README.md file that describes the application and provides instructions on how to install and run it. Also started creating some of the packages and corresponding source code and test files. Starting with the News Source package first as it needs to be implemented fully before we can create the cli package/application. Just started writing the initial files. I will continue to add more functionality and files throughout the week.